rename `chalk_solve::slg::implementation` to `chalk_solve::slg`
